web/src/app/components/property/many-to-many-relationship-property-artifacts/many-to-many-relationship-property-artifacts.component.ts
import { Component, OnInit, Input } from '@angular/core';

import { ArtifactsManyToManyRelationshipPropertyModel } from '../../../artifacts.model';

@Component({
  selector: 'app-many-to-many-relationship-property-artifacts',
  templateUrl: './many-to-many-relationship-property-artifacts.component.html',
  styleUrls: ['./many-to-many-relationship-property-artifacts.component.css'],
})
export class ManyToManyRelationshipPropertyArtifactsComponent
  implements OnInit {
  @Input() artifacts: ArtifactsManyToManyRelationshipPropertyModel;

  constructor() {}

  ngOnInit(): void {}
}
